+++ src/condition.cpp
@@ -1602,7 +1602,7 @@ std::function<double( dialogue const & )> conditional_t::get_get_dbl( J const &j
         } else if( checked_value == "skill_level" ) {
             const skill_id skill( jo.get_string( "skill" ) );
             return [is_npc, skill]( dialogue const & d ) {
-                return d.actor( is_npc )->get_skill_level( skill );
+                return static_cast<int>( d.actor( is_npc )->get_skill_level( skill ) );
             };
         } else if( checked_value == "pos_x" ) {
             return [is_npc]( dialogue const & d ) {
