+++ src/trap.cpp
@@ -250,7 +250,7 @@ bool trap::detect_trap( const tripoint &pos, const Character &p ) const
     // Obviously it rapidly gets better as your skills improve.
 
     // Devices skill is helpful for spotting traps
-    const int traps_skill_level = p.get_skill_level( skill_traps );
+    const float traps_skill_level = p.get_skill_level( skill_traps );
 
     // Perception is the main stat for spotting traps, int helps a bit.
     // In this case, stats are more important than skills.
